notice section fully optimzed
